# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kanban API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
        "/boards": {
            "get": {
                "tags": [
                    "Boards"
                ],
                "summary": "Obtener todos los tableros",
                "operationId": "80c5f0b927bb08b0075ffec38d6e2c4a",
                "responses": {
                    "200": {
                        "description": "Lista de tableros obtenida exitosamente"
                    }
                }
            },
            "post": {
                "tags": [
                    "Boards"
                ],
                "summary": "Crear un nuevo tablero",
                "operationId": "e1f8712c406dca13de55015c113ad641",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Mi Proyecto"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción del proyecto"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tablero creado exitosamente"
                    }
                }
            }
        },
        "/boards/{id}": {
            "get": {
                "tags": [
                    "Boards"
                ],
                "summary": "Obtener un tablero específico",
                "operationId": "7cfdbcedb97c27684bac56171172fffe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tablero",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tablero obtenido exitosamente"
                    },
                    "404": {
                        "description": "Tablero no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Boards"
                ],
                "summary": "Actualizar un tablero",
                "operationId": "02db9041d873012a0cdee77b129c1cb6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tablero",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nombre Actualizado"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tablero actualizado exitosamente"
                    },
                    "404": {
                        "description": "Tablero no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Boards"
                ],
                "summary": "Eliminar un tablero",
                "operationId": "4296320124bd00d1551152ec505f3635",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tablero",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tablero eliminado exitosamente"
                    },
                    "404": {
                        "description": "Tablero no encontrado"
                    }
                }
            }
        },
        "/boards/{id}/tasks": {
            "get": {
                "tags": [
                    "Boards"
                ],
                "summary": "Obtener tareas de un tablero con filtros",
                "operationId": "4aa92200be7e36626dc6a6ee632360fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tablero",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "prioridad",
                        "in": "query",
                        "description": "Filtrar por prioridad",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "alta",
                                "media",
                                "baja"
                            ]
                        }
                    },
                    {
                        "name": "porcentaje",
                        "in": "query",
                        "description": "Filtrar por porcentaje de progreso",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "vencimiento",
                        "in": "query",
                        "description": "Filtrar por fecha de vencimiento (YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "buscar",
                        "in": "query",
                        "description": "Buscar en título y descripción",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tareas obtenidas exitosamente"
                    },
                    "404": {
                        "description": "Tablero no encontrado"
                    }
                }
            }
        },
...
    },
...
}